TP3/pyproject.toml
[tool.poetry]
name = "Tp3-Computer-Vision"
version = "0.1.0"
description = ""
authors = ["Nicolas Depelteau <nicolas@depelteau.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
notebook = "^7.0.6"
ipykernel = "^6.29.0"
scikit-image = "^0.22.0"
matplotlib = "^3.8.2"
tqdm = "^4.66.1"
python-dotenv = "^1.0.0"
wandb = "^0.16.2"
hydra-core = "^1.3.2"
scikit-learn = "^1.4.0"
hydra-joblib-launcher = "^1.2.0"
pandas = "^2.2.0"
poethepoet = "^0.24.4"
torchmetrics = "^1.3.1"

[tool.poe.tasks]
install-pytorch = "pip install torch torchaudio torchvision --index-url https://download.pytorch.org/whl/cu121"
install-torchvision = "pip install torchvision --no-deps --force-reinstall --index-url https://download.pytorch.org/whl/cu121"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
